# swagger.yaml
openapi: 3.0.3
info:
  title: Dating App API
  version: 1.0.0
  description: |
    Music-based dating application backend. Matches users based on Spotify listening history,
    lifestyle compatibility, and behavioral learning.

    ## Authentication
    Most endpoints require a JWT passed as `Authorization: Bearer <token>`.
    JWT claims: `{ userId, tokenVersion, sub, iat, exp }`.

    After a password reset `tokenVersion` is incremented — all prior tokens are invalidated.
    Deleted-user JWTs return **401** (not 404).

    ## Email Verification
    Unverified-email users receive **403** `{"error":"Email verification required"}` on all
    endpoints except `/api/v1/auth/**`, `/api/v1/onboarding/**`, `/public/**`, and `/api/test/**`.

    ## Rate Limiting
    | Action | Limit | Key |
    |--------|-------|-----|
    | Swipe | 60/min | userId |
    | Score calculation | 30/min | userId |
    | Unauthenticated | per IP | remoteAddr |

    Exceeded limits return **429**.

    ## Error Contract
    Every error response follows:
    ```json
    { "code": "string", "message": "string", "fields": {}, "timestamp": "ISO8601" }
    ```
    Always switch on `code`, never on `message`.

  contact:
    name: API Support
    email: mladen.hangi@kodelab.hr
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/preferences/genres:
    get:
      tags: [Genre Preferences]
      summary: Get current user's genre preferences
      operationId: getGenrePreferences
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Genre preferences list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenrePreferenceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Genre Preferences]
      summary: Add a manual genre preference
      operationId: addGenrePreference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGenrePreferenceRequest'
            example:
              genreName: indie-rock
              weight: 0.9
      responses:
        '201':
          description: Genre preference added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenrePreferenceCreateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/preferences/genres/sync:
    post:
      tags: [Genre Preferences]
      summary: Sync genre preferences from Spotify
      operationId: syncGenrePreferences
      parameters:
        - name: quick
          in: query
          schema:
            type: boolean
            default: false
          description: Quick sync uses only short-term listening data
      responses:
        '200':
          description: Sync successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreSyncResponse'
              example:
                success: true
                message: Genre preferences synced successfully
                genreCount: 12
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/SpotifyTokenExpired'
        '502':
          $ref: '#/components/responses/SpotifyUnavailable'
  /api/v1/preferences/genres/spotify:
    delete:
      tags: [Genre Preferences]
      summary: Clear all Spotify-derived genre preferences
      operationId: clearSpotifyGenrePreferences
      responses:
        '200':
          description: Spotify preferences cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
              example:
                success: true
                message: Spotify preferences cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/preferences/genres/{genreName}:
    delete:
      tags: [Genre Preferences]
      summary: Remove a specific genre preference
      operationId: removeGenrePreference
      parameters:
        - name: genreName
          in: path
          required: true
          schema:
            type: string
          example: indie-rock
      responses:
        '200':
          description: Genre preference removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
              example:
                success: true
                message: Genre preference removed
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── Account ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT with claims: `{ userId, tokenVersion, sub, iat, exp }`.
        `tokenVersion` is required — tokens without it are rejected with 401.

  # ─── Responses ─────────────────────────────────────────────────────────────
  responses:
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            fields:
              email: must not be blank
              password: Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character
            timestamp: '2026-05-12T10:15:30.123Z'
    InvalidArgument:
      description: Invalid query parameter or illegal request state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_ARGUMENT
            message: Invalid status parameter. Allowed values active, all
            timestamp: '2026-05-12T10:15:30.123Z'
...
    Unauthorized:
      description: Missing or invalid JWT, or wrong credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid email or password
            timestamp: '2026-05-12T10:15:30.123Z'
...
    SpotifyTokenExpired:
      description: Spotify refresh token revoked — user must re-authenticate with Spotify
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: SPOTIFY_TOKEN_EXPIRED
            message: Spotify authorization has been revoked. Please re-connect your Spotify account.
            timestamp: '2026-05-12T10:15:30.123Z'
    SpotifyUnavailable:
      description: Spotify API error or circuit breaker open
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: SPOTIFY_UNAVAILABLE
            message: Spotify service is temporarily unavailable
            timestamp: '2026-05-12T10:15:30.123Z'
...
  schemas:
...
    GenrePreferenceSource:
      type: string
      description: Serialized as lowercase in JSON
      enum: [spotify_derived, manual_selection, inferred, hybrid, seed_data]
...
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: Machine-readable error code. Always switch on this, never on message.
          example: VALIDATION_ERROR
          enum:
            - VALIDATION_ERROR
            - INVALID_ARGUMENT
            - INVALID_TOKEN
            - UNAUTHORIZED
            - FORBIDDEN
            - ACCOUNT_LOCKED
            - NOT_FOUND
            - EMAIL_EXISTS
            - CONFLICT
            - DUPLICATE_SWIPE
            - CONCURRENT_MODIFICATION
            - SPOTIFY_TOKEN_EXPIRED
            - SPOTIFY_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable description
        fields:
          type: object
          nullable: true
          description: Per-field validation messages. Only present for VALIDATION_ERROR.
          additionalProperties:
            type: string
          example:
            email: must not be blank
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp

    # ── Auth ───────────────────────────────────────────────────────────────
...
    GenrePreferenceItem:
      type: object
      properties:
        genreName:
          type: string
          example: indie-rock
        genreDisplayName:
          type: string
          example: Indie Rock
        weight:
          type: number
          format: double
          description: 0.0–1.0
          example: 0.85
        confidence:
          type: number
          format: double
          example: 0.9
        source:
          $ref: '#/components/schemas/GenrePreferenceSource'
        rank:
          type: integer
          example: 1
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    GenrePreferenceListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 12
        preferences:
          type: array
          items:
            $ref: '#/components/schemas/GenrePreferenceItem'
    AddGenrePreferenceRequest:
      type: object
      required: [genreName]
      properties:
        genreName:
          type: string
          example: indie-rock
        weight:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
          default: 1.0
          nullable: true
    GenrePreferenceCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Genre preference added
        preference:
          $ref: '#/components/schemas/GenrePreferenceItem'
    GenreSyncResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Genre preferences synced successfully
        genreCount:
          type: integer
          example: 12

    # ── Account / Users ────────────────────────────────────────────────────
...
    SuccessMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string